[Servers] Convert all servers logos in PNG format
Since version 2.42.11 of gdk-pixbuf, fringe loaders are disabled by 
default.

Cf. 
https://gitlab.gnome.org/GNOME/gdk-pixbuf/-/merge_requests/163/diffs?commit_id=e052a112075a19fb75f1f2ff3de4c82923de13f2

The current version of the GNOME SDK is not built with the `ico` 
(image/x-icon) loader. As we don't know if it will be restored, or when 
if it's an error, it's best to abandon this image format for good. The 
new file extension is now `.png`. Accepted/recommanded formats are PNG, 
JPEG, SVG, WEBP.