[Servers] Continuously updated modules
Add support for external modules

The first use of this new feature is the ability to separately install
an alternative version of the server modules, from the source code
repository, which is updated at each startup and used instead of those
delivered with the application.

The intention is to be able to use the most up-to-date versions of
servers modules, including new servers, without waiting for a new
version of the application.