[Codeberg] Make hardware question multiple selection and make Code of Conduct agreement visible in issue template
A user can use this app on multiple hardware devices so they should be
able to select multiple options.

The Code of Conduct field still appears, even when the options are only
visible to the form. To make it more cohesive, the checkbox to agree to
adhere to the Code of Conduct will remain in the issue content.