[Webview] Improve BypassCF decorator
New challenge detection method (presence/absence and resolution)

This is now done at the COMMITTED loading event. It's simpler and 
faster.

In case server don't used Cloudflare (temporarily or not at all), a fake 
`cf_clearance` cookie is now created, so as not to try to detect the 
challenge next time.